Former-commit-id: 71b24ef85c10d7860c19602c89823a6472a48283

+++ sql/tool/docker/cajamar/install.sh
@@ -4,6 +4,8 @@ CURRENT_DUMP_NAME=$1
 STARTING_TAG=$2
 CONTAINER_NAME=$3
 CUSTOM_NLS_LANG=$4
+OPTION_RANDOM_DUMP=$5
+OPTION_REMOVE=$6
 
 if [[ "x$CURRENT_DUMP_NAME" == "x" || "x$STARTING_TAG" == "x" || "x$CONTAINER_NAME" == "x" || "x$CUSTOM_NLS_LANG" == "x" ]]; then
 	echo "ERROR: No se puede continuar con la instalación de la BBDD"
@@ -18,7 +20,7 @@ INNER_DUMP_DIRECTORY=/DUMP
 # OUTSIDE DOCKER
 DUMP_FILE_OUT_DOCKER=DUMP/$CURRENT_DUMP_NAME
 if [[ "x$(hostname)" != "x$CONTAINER_NAME" ]]; then
-	docker exec $CONTAINER_NAME /setup/install.sh $CURRENT_DUMP_NAME $STARTING_TAG $CONTAINER_NAME $CUSTOM_NLS_LANG
+	docker exec $CONTAINER_NAME /setup/install.sh $CURRENT_DUMP_NAME $STARTING_TAG $CONTAINER_NAME $CUSTOM_NLS_LANG $OPTION_RANDOM_DUMP $OPTION_REMOVE
 	exit 0
 fi
 
@@ -51,30 +53,39 @@ echo "<Docker [$CONTAINER_NAME]>: BBDD disponible: OK"
 DUMP_FILE_PATH=$INNER_DUMP_DIRECTORY/$CURRENT_DUMP_NAME
 
 if [[ -f $DUMP_FILE_PATH  ]]; then
-	echo "<Docker [$CONTAINER_NAME]>: Limpiando el contenido de /oradata..."
-	rm -f /oradata/*
-	echo "<Docker [$CONTAINER_NAME]>: creando tablespaces y directorios..."
-	$ORACLE_HOME/bin/sqlplus system/admin@localhost:1521/orcl @/setup/SQL-SCRIPTS/script.sql
-	chmod go+rw /oradata/*
+	if [[ "x$OPTION_REMOVE" == "xyes" ]]; then
+		echo "<Docker [$CONTAINER_NAME]>: Limpiando el contenido de /oradata..."
+		rm -f /oradata/*
+		$ORACLE_HOME/bin/sqlplus system/admin@localhost:1521/orcl @/setup/SQL-SCRIPTS/alter-system-user.sql &>/dev/null
+		echo "<Docker [$CONTAINER_NAME]>: creando tablespaces y directorios..."
+		$ORACLE_HOME/bin/sqlplus system/admin@localhost:1521/orcl @/setup/SQL-SCRIPTS/script.sql
+		chmod go+rw /oradata/*
+	fi
 	
 	echo "<Docker [$CONTAINER_NAME]>: Importando dump de la bbdd.."
 	$ORACLE_HOME/bin/impdp system/admin@localhost:1521/orcl DIRECTORY=scripts dumpfile=$CURRENT_DUMP_NAME logfile=SYSTMP:$CURRENT_DUMP_NAME.import.log schemas=CM01,CMMASTER remap_tablespace=BANK01:DRECOVERYONL8M,TEMPORAL:TEMP
-	
-	echo "<Docker [$CONTAINER_NAME]>: Asignando passwords a usuarios de Cajamar..."
-	$ORACLE_HOME/bin/sqlplus system/admin@localhost:1521/orcl @/setup/SQL-SCRIPTS/cmpasswords.sql
-	
-	echo "<Docker [$CONTAINER_NAME]>: Configurando ENTIDADCONFIG..."
-	$ORACLE_HOME/bin/sqlplus system/admin@localhost:1521/orcl @/setup/SQL-SCRIPTS/DML-00-CMMASTER-update.sql
 
-	echo "<Docker [$CONTAINER_NAME]>: Ejecutando scripts..."
-	export PATH=$PATH:$ORACLE_HOME/bin
-	export NLS_LANG=$CUSTOM_NLS_LANG
-	echo "export NLS_LANG=$NLS_LANG" >> /home/oracle/.bashrc
-	echo "<Docker [$CONTAINER_NAME]>: NLS_LANG=$NLS_LANG"
-	cd /sql-package/DDL
-	./DDL-scripts.sh admin@orcl admin@orcl
-	cd /sql-package/DML
-	./DML-scripts.sh admin@orcl admin@orcl
+	if [[ "x$OPTION_REMOVE" == "xyes" ]]; then
+		echo "<Docker [$CONTAINER_NAME]>: Asignando passwords a usuarios de Cajamar..."
+		$ORACLE_HOME/bin/sqlplus system/admin@localhost:1521/orcl @/setup/SQL-SCRIPTS/cmpasswords.sql
+
+		echo "<Docker [$CONTAINER_NAME]>: Configurando ENTIDADCONFIG..."
+		$ORACLE_HOME/bin/sqlplus system/admin@localhost:1521/orcl @/setup/SQL-SCRIPTS/DML-00-CMMASTER-update.sql
+	fi
+
+	if [[ "x$OPTION_RANDOM_DUMP" != "xyes" ]]; then
+		echo "<Docker [$CONTAINER_NAME]>: Ejecutando scripts..."
+		export PATH=$PATH:$ORACLE_HOME/bin
+		export NLS_LANG=$CUSTOM_NLS_LANG
+		echo "export NLS_LANG=$NLS_LANG" >> /home/oracle/.bashrc
+		echo "<Docker [$CONTAINER_NAME]>: NLS_LANG=$NLS_LANG"
+		cd /sql-package/DDL
+		./DDL-scripts.sh admin@orcl admin@orcl
+		cd /sql-package/DML
+		./DML-scripts.sh admin@orcl admin@orcl
+	else
+		echo "<Docker [$CONTAINER_NAME]>: No se van a ejecutar scripts DDL ni DML"
+	fi
 else
 	echo "*************************** WARNING *******************************"
 	echo "<Docker [$CONTAINER_NAME]>: $DUMP_FILE_PATH: No existe el fichero no se va a importar el DUMP"
